factor::numeric: Move Montgomery arithmetic to its own module
Finally hollowed-out numeric.rs
